+++ src/main/java/subway/auth/infrastructure/AuthenticationPrincipalArgumentResolver.java
@@ -1,4 +1,4 @@
-package subway.auth.ui;
+package subway.auth.infrastructure;
 
 import org.springframework.core.MethodParameter;
 import org.springframework.web.bind.support.WebDataBinderFactory;
