+++ src/main/java/subway/common/ExceptionAdvice.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.common;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
